requirements.txt
diffusers
transformers
matplotlib
pyrallis
omegaconf
opencv-python
einops
pytorch-lightning
accelerate
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16